+++ R/f_val_calc_all_driver.R
@@ -143,6 +143,10 @@ function# This function calls all the \code{f_val_calc.*} files
   # val_TDL   <- val$sum$TDL;
   # val_Licor <- val$sum$Licor;
   # val_const <- val$const;
+  #
+  # val_TDL   <- val$obs$TDL;
+  # val_Licor <- val$obs$Licor;
+  # val_const <- val$const;
   # DEBUG DEBUG DEBUG DEBUG DEBUG
   val_temp <- list(); # hold current set of calculated values below